Move `context_plugin` out of `plugin.hpp`
