+++ libtenzir/builtins/contexts/bloom_filter.cpp
@@ -10,6 +10,7 @@
 #include <tenzir/concept/parseable/numeric.hpp>
 #include <tenzir/concept/parseable/tenzir/expression.hpp>
 #include <tenzir/concept/parseable/to.hpp>
+#include <tenzir/context.hpp>
 #include <tenzir/data.hpp>
 #include <tenzir/dcso_bloom_filter.hpp>
 #include <tenzir/detail/range_map.hpp>
@@ -18,7 +19,6 @@
 #include <tenzir/flatbuffer.hpp>
 #include <tenzir/fwd.hpp>
 #include <tenzir/operator.hpp>
-#include <tenzir/plugin.hpp>
 #include <tenzir/series.hpp>
 #include <tenzir/series_builder.hpp>
 #include <tenzir/table_slice.hpp>
