+++ libtenzir/src/detail/add_message_types.cpp
@@ -17,6 +17,7 @@
 #include "tenzir/component_registry.hpp"
 #include "tenzir/config.hpp"
 #include "tenzir/connect_request.hpp"
+#include "tenzir/context.hpp"
 #include "tenzir/detail/stable_map.hpp"
 #include "tenzir/die.hpp"
 #include "tenzir/expression.hpp"
@@ -70,9 +71,11 @@ void add_message_types() {
   // Check for type ID conflicts between dynamic plugins.
   for (const auto& [new_block, assigner] :
        plugins::get_static_type_id_blocks()) {
-    for (const auto& old_block : old_blocks)
-      if (new_block.begin < old_block.end && old_block.begin < new_block.end)
+    for (const auto& old_block : old_blocks) {
+      if (new_block.begin < old_block.end && old_block.begin < new_block.end) {
         die("cannot assign overlapping plugin type ID blocks");
+      }
+    }
     old_blocks.push_back(new_block);
     assigner();
   }
